+++ changelog.md
@@ -17,5 +17,6 @@
 - FIX: FilesTree folder expanding add itesm again after collapse and expand.
 - ADD: FilesTree support to copy 'n paste a file.
 - ADD: FilesTree support to cut 'n paste a file.
-- ADD: popup Menu with a close button to the CMD box."
+- ADD: popup Menu with a close button to the CMD box."a
+- ADD: show current git branch in status line.